test: migrate sanity tests from legacy format to Vitest with comprehensive checks
